+++ pAudio/code/share/config.py
@@ -174,7 +174,7 @@ def _init():
 
                         fir_path = f'{LSPKFOLDER}/{fs}/drc.{ch}.{set_name}.pcm'
 
-                        channels[ch]["1"] = make_fir_filter(fir_path)
+                        channels[ch][1] = make_fir_filter(fir_path)
 
                     LSPK_CONFIG["drc"][set_name] = channels
 
@@ -386,6 +386,9 @@ def _init():
 
     if not 'ref_level_gain_offset' in CONFIG:
         CONFIG["ref_level_gain_offset"] = 0.0
+
+    if not 'loudness_compensation_clamped_above_zero' in CONFIG:
+        CONFIG["loudness_compensation_above_zero"] = False
 
     if not "tones_span_dB" in CONFIG:
         CONFIG["tones_span_dB"] = 6.0