+++ pAudio/start.py
@@ -83,7 +83,6 @@ def prepare_jack_stuff():
     if any('mpd' in p for p in CONFIG["plugins"]):
         jloops_list.append('mpd_loop')
 
-
     fs       = CONFIG["samplerate"]
     alsa_dev = CONFIG["jack"]["device"]
     period   = CONFIG["jack"]["period"]
@@ -331,6 +330,9 @@ def stop():
 
 def start():
 
+    # restore Sound Card settings (currently only for Linux-ALSA)
+    restore_sound_card()
+
     # Check if CamillaDPS is available
     if not check_cdsp_running():
         return
@@ -395,7 +397,6 @@ if __name__ == "__main__":
 
         case 'stop':
             stop()
-
 
         case 'prepare_jack_stuff':
             prepare_jack_stuff()