+++ doc/NEWS.md
@@ -8,3 +8,8 @@
 ## 2024-jun: MacOS Coreaudio and Linux JACK
 
 - JACK sound server has been implemented for Linux systems, it allows for preamp "input" selector features and more, for instance multiroom JACK to JACK systems.
+
+## 2025-dec: players management
+
+- new code for players management: librespot (Spotify Connect client)
+- minor fixes